+++ src/main/groovy/com/episode6/hackit/deployable/DeployableJarPlugin.groovy
@@ -24,7 +24,7 @@ class DeployableJarPlugin implements Plugin<Project> {
     }
 
     project.deployable {
-      primaryPublication {
+      mainArtifact {
         artifact project.jar
       }
       publication {