+++ src/main/groovy/com/episode6/hackit/deployable/DeployableKotlinJarPlugin.groovy
@@ -35,7 +35,7 @@ class DeployableKotlinJarPlugin implements Plugin<Project> {
     }
 
     project.deployable {
-      primaryPublication {
+      mainArtifact {
         artifact project.jar
       }
       publication {