+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### v0.2.0-SNAPSHOT - unreleased
 - Re-write plugin to use `maven-publish` instead of old `maven` plugin
     - Api stays mostly the same, deploy and install tasks are still valid
-- **[breaking]** Customizing deployable artifacts has changed, we now use the `deployable.publication { artifact task }` and `deployable.primaryPublication { artifact task }` blocks
+- **[breaking]** Customizing deployable artifacts has changed, we now use the `deployable.publication { artifact task }` and `deployable.mainArtifact { artifact task }` blocks
 - Add config block to customize pom as xml (after deployable has done its initial setup) `deployable.pom.withXml { }`
 - Add new plugin `com.episode6.hackit.deployable.gradle-plugin` to workaround java-gradle-plugins build in publish config
 - **[breaking]** Stop providing default values for repo urls. If no urls are specified, no repo will be set up.